+++ protocols/testing/src/tycho_runner.rs
@@ -333,7 +333,7 @@ pub fn get_default_endpoint(chain: &Chain) -> Option<String> {
         Chain::Base => Some("https://base-mainnet.streamingfast.io:443".to_string()),
         Chain::Unichain => Some("https://mainnet.unichain.streamingfast.io:443".to_string()),
         Chain::Polygon => Some("https://polygon.streamingfast.io:443".to_string()),
-        Chain::Robinhood => Some("https://robinhood.substreams.pinax.network:443".to_string()),
+        Chain::Robinhood => Some("https://mainnet.robinhood.streamingfast.io:443".to_string()),
         _ => None,
     }
 }